Put all files except the main.c file into a directory with a functional name
