+++ include/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: seunan <seunan@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/02 12:13:27 by seunan            #+#    #+#             */
-/*   Updated: 2023/12/11 15:24:19 by seunan           ###   ########.fr       */
+/*   Updated: 2023/12/14 20:22:21 by seunan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
