--- include/matrix.h
@@ -1,19 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   matrix.h                                           :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: sunko <sunko@student.42.fr>                +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/12/11 16:12:37 by sunko             #+#    #+#             */
-/*   Updated: 2023/12/12 11:53:22 by sunko            ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef MATRIX_H
-# define MATRIX_H
-
-t_4x4matrix	_4x4matrix(t_vec4 r1, t_vec4 r2, t_vec4 r3, t_vec4 r4);
-t_vec3		mv_mul(t_4x4matrix m, t_vec4 v);
-
-#endif
