+++ src/print/color.c
@@ -6,12 +6,11 @@
 /*   By: seunan <seunan@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 12:48:22 by seunan            #+#    #+#             */
-/*   Updated: 2023/12/13 22:20:58 by seunan           ###   ########.fr       */
+/*   Updated: 2023/12/14 20:24:59 by seunan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "structures.h"
-#include "minirt.h"
 
 t_color3	get_color_int_to_real(t_color3 i_rgb)
 {
